airflow-docker/requirements.txt
apache-airflow-providers-google
google-cloud-storage
google-cloud-bigquery
requests
pandas